+++ VSMS.Application/Components/Shared/Grids/UsersGrid.razor.cs
@@ -190,6 +190,9 @@ public partial class UsersGrid : ComponentBase
     {
         try
         {
+            var isConfirmed = await ShowConfirmModal();
+            if (!isConfirmed) return;
+            
             var res = await UsersHttpService.DeleteUser(userId);
             if (res)
                 await _dataGrid.ReloadServerData();
@@ -199,6 +202,40 @@ public partial class UsersGrid : ComponentBase
         catch (Exception e)
         {
             Logger.LogError(e, e.Message);
+        }
+    }
+    
+    private async Task<bool> ShowConfirmModal()
+    {
+        try
+        {
+            var parameters = new DialogParameters<ConfirmationModal>
+            {
+                { x => x.Message, $"user_delete_confirmation_message" },
+            };
+            
+            var options = new DialogOptions
+            {
+                Position = DialogPosition.Center,
+                CloseOnEscapeKey = true,
+                NoHeader = false,
+                CloseButton = true,
+                FullScreen = false,
+                FullWidth = false,
+                MaxWidth = MaxWidth.Large,
+            };
+            
+            var dialogReference = await DialogService.ShowAsync<ConfirmationModal>("", parameters, options);
+            var dialogResult = await dialogReference.Result;
+            if (dialogResult is { Canceled: true })
+                return false;
+            
+            return dialogResult?.Data is bool;
+        }
+        catch (Exception e)
+        {
+            Logger.LogError(e, e.Message);
+            return false;
         }
     }
 }
